# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin OpenApi",
        "description": "this is a api documentation",
        "contact": {
            "name": "Emad shirzad",
            "email": "shemad2022@gmail.com"
        },
        "license": {
            "name": "GPL-v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/auth/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "register",
                "description": "register",
                "operationId": "register",
                "requestBody": {
                    "description": "tasks input",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "username ",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "password confirm",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "login",
                "description": "login",
                "operationId": "login",
                "requestBody": {
                    "description": "tasks input",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "username",
                                        "type": "string",
                                        "example": "Test User"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user/role/{id}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "changeRole",
                "description": "changeRole (Your must be admin)",
                "operationId": "changeRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "tasks input",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "description": "role",
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "phone": {
                                        "description": "phone (optional)",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "description": "address (optional)",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "balance": {
                                        "description": "balance (optional)",
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/auth/user/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "my info",
                "description": "my info",
                "operationId": "48f639b92fda544efac156d3a00cc179",
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/auth/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "logout",
                "description": "logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/auth/user/refresh": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "refresh",
                "description": "refresh a token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessModel": {
                "title": "Success Model",
                "description": "Success Model",
                "properties": {
                    "message": {
                        "description": "Success message",
                        "type": "string",
                        "format": "",
                        "example": "my message"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "title": "Error Model",
                "description": "Error Model",
                "properties": {
                    "message": {
                        "description": "error message",
                        "type": "string",
                        "format": "",
                        "example": "my error message"
                    },
                    "errors": {
                        "description": "errors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "error1",
                            "error2"
                        ]
                    }
                },
                "type": "object"
            },
...
            "UserModel": {
                "title": "User Model",
                "description": "Represents a user",
                "properties": {
                    "id": {
                        "description": "user ID",
                        "type": "integer",
                        "format": "int32"
                    },
                    "username": {
                        "description": "username",
                        "type": "string"
                    },
                    "password": {
                        "description": "password",
                        "type": "string"
                    },
                    "created_by": {
                        "description": "created by envoy",
                        "type": "string"
                    },
                    "is_ban": {
                        "description": "is_ban",
                        "type": "boolean"
                    },
                    "subscription_duration": {
                        "description": "subscription_duration",
                        "type": "integer"
                    },
                    "last_purchase_date": {
                        "description": "last purchase date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "created date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "updated date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "http",
                "description": "Login with username and password to get the authentication token <div>Example: Bearer token</div>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}